wtf-dashboard-retro/requirements.txt
textual>=0.89.0
yfinance
psutil
httpx
feedparser
winmedia-controller
pycaw
comtypes
wmi
winrt-Windows.Media.Control
winrt-Windows.Foundation
winrt-Windows.Foundation.Collections
winrt-Windows.Storage.Streams
Pillow
tzdata
